Move some top level API classes to a player package
