+++ src/api/java/ladysnake/requiem/api/v1/player/MobResurrectable.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses>.
  */
-package ladysnake.requiem.api.v1;
+package ladysnake.requiem.api.v1.player;
 
 import net.minecraft.entity.mob.MobEntity;
