+++ src/main/java/ladysnake/requiem/common/impl/movement/PlayerMovementAlterer.java
@@ -17,9 +17,9 @@
  */
 package ladysnake.requiem.common.impl.movement;
 
-import ladysnake.requiem.api.v1.RequiemPlayer;
 import ladysnake.requiem.api.v1.entity.MovementAlterer;
 import ladysnake.requiem.api.v1.entity.MovementConfig;
+import ladysnake.requiem.api.v1.player.RequiemPlayer;
 import net.minecraft.client.network.packet.PlayerAbilitiesS2CPacket;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
