+++ src/main/java/ladysnake/requiem/mixin/client/gui/hud/InGameHudMixin.java
@@ -18,9 +18,9 @@
 package ladysnake.requiem.mixin.client.gui.hud;
 
 import com.mojang.blaze3d.platform.GlStateManager;
-import ladysnake.requiem.api.v1.RequiemPlayer;
 import ladysnake.requiem.api.v1.event.minecraft.client.CrosshairRenderCallback;
 import ladysnake.requiem.api.v1.event.minecraft.client.HotbarRenderCallback;
+import ladysnake.requiem.api.v1.player.RequiemPlayer;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.DrawableHelper;
 import net.minecraft.client.gui.hud.InGameHud;
