+++ src/main/java/ladysnake/requiem/mixin/client/render/GameRendererMixin.java
@@ -17,8 +17,8 @@
  */
 package ladysnake.requiem.mixin.client.render;
 
-import ladysnake.requiem.api.v1.RequiemPlayer;
 import ladysnake.requiem.api.v1.event.minecraft.client.ApplyCameraTransformsCallback;
+import ladysnake.requiem.api.v1.player.RequiemPlayer;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.render.Camera;
 import net.minecraft.client.render.GameRenderer;
