+++ src/main/java/ladysnake/requiem/mixin/server/network/ServerPlayerInteractionManagerMixin.java
@@ -17,7 +17,7 @@
  */
 package ladysnake.requiem.mixin.server.network;
 
-import ladysnake.requiem.api.v1.RequiemPlayer;
+import ladysnake.requiem.api.v1.player.RequiemPlayer;
 import ladysnake.requiem.api.v1.remnant.RemnantState;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.network.ServerPlayerInteractionManager;
