+++ src/main/resources/mixins.requiem.client.json
@@ -2,6 +2,7 @@
   "required": true,
   "package": "ladysnake.requiem.mixin.client",
   "compatibilityLevel": "JAVA_8",
+  "minVersion": "0.7.11-SNAPSHOT",
   "mixins": [
     "MinecraftClientMixin",
     "entity.ClientEntityMixin",
