+++ src/main/resources/mixins.requiem.common.json
@@ -2,6 +2,7 @@
   "required": true,
   "package": "ladysnake.requiem.mixin",
   "compatibilityLevel": "JAVA_8",
+  "minVersion": "0.7.11-SNAPSHOT",
   "mixins": [
     "advancement.criterion.CriterionsAccessor",
     "block.entity.LecternBlockEntityMixin",
